requirements.txt
streamlit
tavily-python
google-generativeai
python-dotenv